Continued refactoring... still incomplete
